Small changes to CanvasController.
Created Timestamper.js, the purpose of this commit, pretty sure the whole object is working, checked the chunk-interval with as fast as 20ms and seems to be holding steady, which indicates it's not too CPU intensive due to my previous knowledge of JS's requestAnimationFrame. Also tested 1000ms as interval, which results in perfect sync with my phone clock. Some test still remain to be made however, but the class is, if not perfect, very close.

cache manifest because I noticed the script.js was being cached randomly, don't want to have to Ctrl+F5 everytime I save the file. Didn't test to see if it worked.