ClassTemplate created so that I can easily create new classes and organize them neatly.
CanvasController now has a objects property which is an array that will hold objects such as buttons and other drawable things (entire maps will be objects too)

GuiButton was created, but has no drawing as of now, must finish that class on monday.

I'm very happy with how Timestamper works, it is really efficient and well encapsulated.