Cache Problems Fixed
I thought index.html wouldn't ever change so I put it in the cache, but I just realized that everytime I add a new class I must put it at the header, so I removed it from the cache.

Anyways, my first instict is always to double check that the code truly changed, not because javascript is unreliable but because I've fixed more problems with this technique than I care to admit in every language, from C# to Ruby, so cache problems never have an upper hand on me.