GuiBox creation
Alignment made positioning too complex, so I created another class to handle it: GuiBox, which is in itself a boundingbox with values like Left, Right, Bottom and Top, and two extra helping hands for alignment: Center and Middle, all of which can be assigned to help control the button according to what's necessary. This is a very useful class, one of which I might use frequently in the future.

GuiButton now only handles button-like stuff, like caption, drawing, states and mouse events, which is ideal.

There is still an error that forces width to be zero for some reason, I will focus on that when I have time.